Combined NotesList and NotesAdder into NotesListEditor

+++ Source/WebApps/C8S.AdminApp.Client.UI/Notes/NoteDisplayer.razor.cs
@@ -1,7 +1,7 @@
 ﻿using C8S.AdminApp.Client.Services.Coordinators.Notes;
 using C8S.Domain.Features.Notes.Models;
-using MediatR;
 using Microsoft.AspNetCore.Components;
+using Radzen;
 using SC.Common.Razor.Base;
 
 namespace C8S.AdminApp.Client.UI.Notes;
@@ -9,16 +9,22 @@ namespace C8S.AdminApp.Client.UI.Notes;
 public partial class NoteDisplayer : BaseRazorComponent
 {
     [Inject]
-    public ISender Sender { get; set; } = null!;
+    public DialogService DialogService { get; set; } = null!;
 
     [Parameter]
-    public NoteDetails Note { get; set; } = null!;
+    public NoteDetails Details { get; set; } = null!;
 
     [Parameter]
-    public NotesListCoordinator Coordinator { get; set; } = null!;
+    public NotesListEditorCoordinator Coordinator { get; set; } = null!;
 
     private async Task DeleteNote(int noteId)
     {
+        if (!(await DialogService
+                .Confirm($"This will permanently delete the note written by '{Details.Author}'" +
+                         $" on {Details.CreatedOn:M/d/yy} at {Details.CreatedOn:h:mm tt}.\r\nIs that okay?", "Confirm Delete?", 
+                new ConfirmOptions() { CancelButtonText = "No, Keep It", OkButtonText = "Yes, Delete It"}) ?? false))
+            return;
+
         await Coordinator.DeleteNote(noteId);
     }
     private async Task EditNote(int noteId)
